partically fixed negative total_found, docs and hits counter on huge indexes
git-svn-id: http://sphinxsearch.googlecode.com/svn/branches/rel20/@3080 8b96e2b9-35c5-2c16-bc47-5122d61876d4

+++ src/sphinx.cpp
@@ -24234,7 +24234,7 @@ CSphQueryResultMeta::CSphQueryResultMeta ()
 }
 
 
-void CSphQueryResultMeta::AddStat ( const CSphString & sWord, int iDocs, int iHits, bool bExpanded )
+void CSphQueryResultMeta::AddStat ( const CSphString & sWord, int64_t iDocs, int64_t iHits, bool bExpanded )
 {
 	CSphString sFixed;
 	const CSphString * pFixed = &sWord;
